Create working code
This code gives the probability of a slip plane intersecting a percipitate in a randon lattice of a material. You can change the radius of the percipitates by changing the "r" value. There will be between zero and 10 percipitates generated each time and the probability will be printed in the terminal.